Add userquery & role migrations
Add:

- userquery.query_role migration. This allows us to specify a user's
mandatory role on a query. This is primarily to assist us in display of
messages (specifically, which side of the display pane a message should
be displayed). Having a specific role on a query may seem extravagant
but I don't think we can assume a user's role would be the same on every
query (as dictated by their role)

- role.type migration. This allows is to "group" roles. The idea is that
there will (probably) only ever be two types, "PUBLIC" & "STAFF", these
are never displayed, they are primarily used in conjunction with
userquery.query_role above to help organise message display