Extend Query
The query is a more elaborate entity than it was originally. This commit
makes the following changes:

- Add a query.initiator column, this is a foreign key to the user table.
It holds the user who originally created the query and will be populated
at the time the query is created.
- Adds additional resolvers that allow the return of data associated
with queries, but not necessarily derived from the query table. These
resolvers are designed for bulk return, so will return data for multiple
queries:
  - initiators - given an array of query IDs, this resolver returns an
array of initiators for those queries
  - participants - given an array of query IDs, this resolver returns
an array of everyone involved in all of the suppied queries
  - latestMessages - given an array of query IDs, this resolver returns
a message object for each query's latest message