Remove non-folder "folders"
"Inbox" (now "Unread") & "All queries" are not really folders. Their
contents are derived from the status of a query's messages, if a query
has unread messages, it will appear in "Unread", otherwise it will
appear in "All queries". As such, we don't need explicit folders for
them.

This commit deletes the folders "Inbox" & "All queries"