Fix schema and resolver for adding query label
Somehow I failed to add querylabel.query_id & querylabel.label_id as a composite primary key. I
noticed when duplicate rows were inserted and the UI had an appropriate
fit regarding iterator keys. This becomes a problem for bulk attachment
of label to query, since in the API call we're saying "add label A to
queries X, Y & Z" where query Y may already have it. So we can't be
discerning about what we want in the API call.

The API controller *could* do the discerning for us but that would
involve looking up the existing labels for each affected query and
preventing an insert if a label already exist on a query. (Something
like, get all queries with this label, then when iterating through the
queries, look to see if if it's in the list.

However, I've decided to tackle it at the simplest & lowest level. I've added the
required composite primary key on querylabel and added a "ON CONFLICT DO
NOTHING" to the insert. Since the INSERT is returning *, we just get
back the existing querylabel row when it already exists.