Add auth related columns and resolvers
This commit adds the necessary columns to ems_user to allow
authentication details to be stored. These are specifically related to
data returned from delegated authorisation services such as OAuth2 and
SAML. It also adds resolvers to allow updating and creating of
users following authentication.

Includes:

- Migration to add columns to ems_user
- Resolver to provide the ability to create or update a user based on
the provider that supplied them and their provider ID