Unseen count maintenance and retrieval
Provides the ability to maintain the unseen counts for query / user
combinations. The unseen counts are modified when certain events occur
on messages, created, deleted & seen.

Also provides the ability to retrieve the appropriate unseen counts for
a give user.

The logic surrounding the maintenance of the unseen count is more
complex than I'd like but I'm not sure how to avoid that, you have to
take into account the initiator of the action (the sender of a message,
for example, shouldn't have their unseen count incremented for that
query), their role (all staff should have their unseen count modified
for a given query), whether they are involved in the query (a non staff
member who is participating in a query)

Includes:

- Functions for maintaining the unseen_count, these are called when
particular operations occur
- Calling the functions above in the create and delete message resolvers
- Functions for retrieving unseen counts