Refactor and fix unseen incrementing bugs
Includes:

- Renamed and refactored incrementUnseenCount & incrementUnseenCounts to
upsertQueryUser & upsertQueryUsers as this more accurately reflects what
they do. They may increment the unseen could, but they may just create a
new row
- Added the creation of a queryuser entry when an upload is added, this
is inline with what happens when a new message is created.
- Modified the default value for queryuser.unseen_count. This column
does not have a default value in the DB, we explicitly set it to 0 when
a new row is created. We used to set it to 1, but this is no longer
appropriate since a queryuser row can be created when a new query is
created.