Encrypt, rather than hash, user email
Hashing is no good for email storage because we actually need to be able
to use it! Therefore, we are now encrypting using a Node implementation
of libsodium. Currently we're storing the key in the ecosystem file,
which is clearly not ideal, we may need to think about passing it in
when starting the instance or similar.

- Use libsodium to encrypt and decrypt a user's email address
- Added scripts/create-sodium-key.js to create a sodium compliant key